[Orc] Deduplicate GDB JIT Interface declarations (NFC)
